Refactor analytics.tsx by extracting chart components
Split the 953-line analytics.tsx into smaller, maintainable components:
- ViewerChart: Viewer trends chart with daily aggregation
- PlatformDistributionChart: Platform distribution bar chart
- StreamTable: Recent streams table
- AnalyticsSkeleton: Loading state skeleton

analytics.tsx is now 383 lines (under 400 target).
